fixed JSON reader tests
